# api/openapi.yaml
openapi: 3.1.0
info:
  title: skali API
  description: |
    Client-facing API of skali. Authentication is email+password with
    optional TOTP two-factor; sessions are opaque bearer tokens. The same
    API serves every client: the web BFF, the `skali` CLI, and future
    native apps.

    There is no registration endpoint: users are created by an instance
    admin (`/v1/users`) or by the server operator (`skalid user create`).
    Every user carries an instance-wide role, `admin` (everything,
    including users, nodes, and instance settings) or `member` (nothing
    until granted). Product access is per project and per environment on
    one ladder of roles, `none < read < deploy < maintain < admin`: a
    project membership is the member's default role on every environment
    of the project, a per-environment cell overrides it for one user, and
    an environment ceiling (`max_role`) caps what members inherit.
    Non-members do not see a project at all (`404`); a locked environment
    (`none`) is listed by name but its contents answer `403 forbidden`
    naming the required role. Project and environment payloads carry the
    caller's `access` so clients can disable and explain without guessing.
    All error responses share the `Error` envelope; clients should branch
    on `error.code`, not on messages.

    Security-critical operations sit behind sudo mode: they require the
    session to have proven the user's identity recently (within the
...

...

...
  version: 0.1.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/environments/{id}/access:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: listEnvironmentAccess
      summary: List the environment's cells
      description: |
        Environment read. The explicit per-user overrides on this
        environment; members without a cell hold their project role,
        capped by the ceiling.
      tags: [access]
      responses:
        "200":
          description: The cells.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentAccessListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/environments/{id}/access/{user}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: user
        in: path
        required: true
        description: The user's id or email address.
        schema:
          type: string
    put:
      operationId: setEnvironmentAccess
      summary: Set a cell
      description: |
        Environment admin, sudo mode. The user must be a member of the
        project (`409` otherwise). Any role including `none` (locked).
      tags: [access]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccessRoleRequest" }
      responses:
        "200":
          description: The cell as stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentAccessResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      operationId: removeEnvironmentAccess
      summary: Remove a cell
      description: Environment admin, sudo mode. The user falls back to their project role.
      tags: [access]
      responses:
        "204":
          description: Cell removed.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token from login / 2FA verify.
  responses:
    BadRequest:
      description: Malformed request body or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid token, wrong credentials, or wrong code.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The caller is authenticated but lacks the role the operation needs
        (`error.code = forbidden`; the message names it, for example
        "maintain on environment production required"), or, on
        sudo-mode-gated operations, recent authentication is required
        (`error.code = reauth_required`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        The resource does not exist, or the caller is not a member of the
        project it belongs to (the two are indistinguishable by design).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource is in a state that does not allow the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    AccessRole:
      type: string
      description: |
        One step of the access ladder, each including everything below it.
        `none` locks an environment (listed by name, contents refused);
        `read` sees everything; `deploy` changes only what code runs
        (promote, rollback, restart, cancel, backup create, direct deploys
        whose definition is unchanged); `maintain` owns the blast radius of
        skali.yml (definition changes, values, restore, exec, credential
        reveal); `admin` manages settings, other users' roles, delete and
        teardown. Project memberships use `read` and up.
      enum: [none, read, deploy, maintain, admin]
...
    Member:
      type: object
      description: One user's role on a project (membership) or on an environment (cell).
      required: [user_id, email, name, role]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role: { $ref: "#/components/schemas/AccessRole" }
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - invalid_credentials
            - invalid_token
            - invalid_code
            - forbidden
            - reauth_required
            - not_found
            - conflict
            - version_conflict
            - invalid_manifest
            - rate_limited
            - node_unreachable
            - no_ready_pod
            - registry_disabled
            - registry_unavailable
            - internal
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
...
    EnvironmentAccessListResponse:
      type: object
      required: [access]
      properties:
        access:
          type: array
          items: { $ref: "#/components/schemas/Member" }
    EnvironmentAccessResponse:
      type: object
      required: [access]
      properties:
        access: { $ref: "#/components/schemas/Member" }
    AccessRoleRequest:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/AccessRole" }
...